% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_classifier.R
\name{analyze_mechanism_dominance}
\alias{analyze_mechanism_dominance}
\title{Analyze mechanism dominance}
\usage{
analyze_mechanism_dominance(chromoanagenesis_result, chr_classification)
}
\arguments{
\item{chromoanagenesis_result}{Chromoanagenesis result object}

\item{chr_classification}{Chromosome classification data frame}
}
\value{
List with dominance analysis
}
\description{
Analyze mechanism dominance
}
\keyword{internal}
