% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotation_utils.R
\name{annotate_chromoanagenesis}
\alias{annotate_chromoanagenesis}
\title{Annotate chromoanagenesis events with gene information}
\usage{
annotate_chromoanagenesis(result, gene_granges, driver_genes = NULL)
}
\arguments{
\item{result}{A chromoanagenesis result object from detect_chromoanagenesis()}

\item{gene_granges}{A GRanges object containing gene coordinates. 
Must have a 'symbol' or 'gene_name' metadata column.}

\item{driver_genes}{Optional character vector of cancer driver gene symbols. 
If NULL, uses get_default_drivers() (OncoKB list).}
}
\value{
A list containing:
  \item{driver_hits}{Detailed table of impacted driver genes with mechanism consistency analysis}
  \item{all_impacted_genes}{Table of all impacted genes}
}
\description{
Maps chromoanagenesis events (Chromothripsis, Chromoplexy, Chromoanasynthesis)
to genes, identifying potentially impacted cancer driver genes.
}
\details{
Incorporates logic to check for "Mechanism Consistency":
- Oncogenes are expected to be Amplified (Chromoanasynthesis)
- Tumor Suppressors (TSGs) are expected to be Disrupted/Lost (Chromothripsis/Chromoplexy)
}
