% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_utils.R
\name{annotate_chromoanagenesis_regions}
\alias{annotate_chromoanagenesis_regions}
\title{Annotate CNV segments with chromoanagenesis status}
\usage{
annotate_chromoanagenesis_regions(cnv_data, chromoanagenesis_result)
}
\arguments{
\item{cnv_data}{CNV data frame}

\item{chromoanagenesis_result}{Chromoanagenesis results}
}
\value{
CNV data with added has_chromoanagenesis column
}
\description{
Marks CNV segments that overlap with chromoanagenesis events.
}
\keyword{internal}
