% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_detection.R
\name{build_adjacency_list_v3}
\alias{build_adjacency_list_v3}
\title{Build adjacency list from edges}
\usage{
build_adjacency_list_v3(nodes, edges)
}
\arguments{
\item{nodes}{All nodes}

\item{edges}{All edges}
}
\value{
Adjacency list
}
\description{
Build adjacency list from edges
}
\keyword{internal}
