% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_statistics.R
\name{calculate_adjacency_pvalue}
\alias{calculate_adjacency_pvalue}
\title{Calculate p-value for breakpoint adjacency}
\usage{
calculate_adjacency_pvalue(distance, mu, genome_size)
}
\arguments{
\item{distance}{Observed distance between breakpoints}

\item{mu}{Local rearrangement rate}

\item{genome_size}{Genome size}
}
\value{
P-value
}
\description{
Calculates the probability that two independent breakpoints would occur
within the observed distance by chance.
}
\keyword{internal}
