% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_detection.R
\name{calculate_chain_complexity_v3}
\alias{calculate_chain_complexity_v3}
\title{Calculate chain complexity (v3)}
\usage{
calculate_chain_complexity_v3(chain, chain_SVs)
}
\description{
Calculate chain complexity (v3)
}
\keyword{internal}
