% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoanasynthesis_detection.R
\name{calculate_chromoanasynthesis_complexity}
\alias{calculate_chromoanasynthesis_complexity}
\title{Calculate chromoanasynthesis complexity score}
\usage{
calculate_chromoanasynthesis_complexity(region, topology)
}
\description{
Calculate chromoanasynthesis complexity score
}
\keyword{internal}
