% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_statistics.R
\name{calculate_genomic_distance}
\alias{calculate_genomic_distance}
\title{Calculate genomic distance between two breakpoints}
\usage{
calculate_genomic_distance(node1, node2, genome_size)
}
\arguments{
\item{node1}{First node (format: "chrom:pos")}

\item{node2}{Second node}

\item{genome_size}{Total genome size}
}
\value{
Genomic distance
}
\description{
For inter-chromosomal breakpoints, uses a normalized distance metric.
}
\keyword{internal}
