% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos_visualization.R
\name{circos_viz}
\alias{circos_viz}
\alias{plot_chromoanagenesis_circos}
\title{Circos Plot Visualization for Chromoanagenesis}
\usage{
plot_chromoanagenesis_circos(
  chromoanagenesis_result,
  SV.sample,
  CNV.sample,
  genome = "hg19",
  highlight_mechanisms = c("chromothripsis", "chromoplexy", "chromoanasynthesis"),
  sample_name = "",
  verbose = TRUE
)
}
\arguments{
\item{chromoanagenesis_result}{Result from detect_chromoanagenesis()}

\item{SV.sample}{Original SV data}

\item{CNV.sample}{Original CNV data}

\item{genome}{Reference genome ("hg19" or "hg38")}

\item{highlight_mechanisms}{Which mechanisms to highlight (default: all)}

\item{sample_name}{Sample name for title}
}
\value{
NULL (plot is created directly)
}
\description{
Functions for creating circos plots to visualize genome-wide
chromoanagenesis events using the circlize package.
}
\keyword{(highlighted)}
\keyword{(inner}
\keyword{(middle}
\keyword{(outer}
\keyword{-}
\keyword{Chromoanagenesis}
\keyword{Chromosome}
\keyword{Copy}
\keyword{Creates}
\keyword{Plot}
\keyword{Structural}
\keyword{a}
\keyword{as}
\keyword{chromoanagenesis}
\keyword{circos}
\keyword{genome-wide}
\keyword{ideogram}
\keyword{internal}
\keyword{links)}
\keyword{number}
\keyword{plot}
\keyword{regions}
\keyword{showing:}
\keyword{track)}
\keyword{variants}
\keyword{variations}
