% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_classifier.R
\name{classify_sample_level}
\alias{classify_sample_level}
\title{Sample-level integrated classification}
\usage{
classify_sample_level(chromoanagenesis_result, chr_classification, overlaps)
}
\arguments{
\item{chromoanagenesis_result}{Chromoanagenesis result object}

\item{chr_classification}{Chromosome classification data frame}

\item{overlaps}{Overlap detection results}
}
\value{
List with sample-level classification
}
\description{
Sample-level integrated classification
}
\keyword{internal}
