% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_detection.R
\name{collect_all_breakpoints_v3}
\alias{collect_all_breakpoints_v3}
\title{Collect all breakpoints from SVs}
\usage{
collect_all_breakpoints_v3(inter_chr_SVs)
}
\arguments{
\item{inter_chr_SVs}{Inter-chromosomal SVs}
}
\value{
Data frame of breakpoints with metadata
}
\description{
Collect all breakpoints from SVs
}
\keyword{internal}
