% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_comparison.R
\name{compare_results}
\alias{compare_results}
\title{Compare two chromoplexy detection results}
\usage{
compare_results(result_cf, result_oi, verbose = TRUE)
}
\description{
Compare two chromoplexy detection results
}
\keyword{internal}
