% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoanasynthesis_visualization.R
\name{create_chromoanasynthesis_region_plot}
\alias{create_chromoanasynthesis_region_plot}
\title{Create a single chromoanasynthesis region plot}
\usage{
create_chromoanasynthesis_region_plot(region_detail, sample_name)
}
\arguments{
\item{region_detail}{Region detail object}

\item{sample_name}{Sample name}
}
\value{
ggplot object
}
\description{
Create a single chromoanasynthesis region plot
}
\keyword{internal}
