% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_visualization.R
\name{create_chromoplexy_chain_plot}
\alias{create_chromoplexy_chain_plot}
\title{Create a single chromoplexy chain plot}
\usage{
create_chromoplexy_chain_plot(
  chain,
  chain_SVs,
  chr_lengths,
  sample_name,
  show_cn,
  CNV.sample
)
}
\arguments{
\item{chain}{Chain object}

\item{chain_SVs}{SVs in chain}

\item{chr_lengths}{Chromosome length data}

\item{sample_name}{Sample name}

\item{show_cn}{Show copy number}

\item{CNV.sample}{CNV data}
}
\value{
ggplot object
}
\description{
Create a single chromoplexy chain plot
}
\keyword{internal}
