% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_comparison.R
\name{create_comparison_summary_df}
\alias{create_comparison_summary_df}
\title{Create comparison summary data frame for multiple samples}
\usage{
create_comparison_summary_df(comparison_results, sample_ids)
}
\arguments{
\item{comparison_results}{List of comparison results from multiple samples}

\item{sample_ids}{Vector of sample IDs}
}
\value{
Data frame with comparison summary
}
\description{
Create comparison summary data frame for multiple samples
}
