% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoanasynthesis_detection.R
\name{create_empty_chromoanasynthesis_result}
\alias{create_empty_chromoanasynthesis_result}
\title{Create empty chromoanasynthesis result}
\usage{
create_empty_chromoanasynthesis_result()
}
\value{
Empty result structure
}
\description{
Create empty chromoanasynthesis result
}
\keyword{internal}
