% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_statistics.R
\name{create_empty_chromoplexy_result_v2}
\alias{create_empty_chromoplexy_result_v2}
\title{Create empty chromoplexy result (v2)}
\usage{
create_empty_chromoplexy_result_v2()
}
\value{
Empty result structure
}
\description{
Create empty chromoplexy result (v2)
}
\keyword{internal}
