% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_detection.R
\name{create_empty_chromoplexy_result_v3}
\alias{create_empty_chromoplexy_result_v3}
\title{Create empty result (v3)}
\usage{
create_empty_chromoplexy_result_v3()
}
\description{
Create empty result (v3)
}
\keyword{internal}
