% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoanagenesis_integrated.R
\name{create_integrated_summary}
\alias{create_integrated_summary}
\title{Create integrated summary of chromoanagenesis results}
\usage{
create_integrated_summary(results)
}
\arguments{
\item{results}{Results list from detect_chromoanagenesis}
}
\value{
Integrated summary data frame
}
\description{
Create integrated summary of chromoanagenesis results
}
\keyword{internal}
