% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_classifier.R
\name{debug_chromoanagenesis_result}
\alias{debug_chromoanagenesis_result}
\title{Debug function to inspect chromoanagenesis results}
\usage{
debug_chromoanagenesis_result(chromoanagenesis_result)
}
\arguments{
\item{chromoanagenesis_result}{Result from detect_chromoanagenesis()}
}
\description{
Debug function to inspect chromoanagenesis results
}
