% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos_debug.R
\name{debug_circos_svs}
\alias{debug_circos_svs}
\title{Debug circos visualization}
\usage{
debug_circos_svs(sv_data)
}
\arguments{
\item{sv_data}{SV data (data frame)}
}
\description{
Helper function to debug circos plot issues
}
