% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_detection.R
\name{deduplicate_chains_v3}
\alias{deduplicate_chains_v3}
\title{Deduplicate chains (v3)}
\usage{
deduplicate_chains_v3(chains)
}
\arguments{
\item{chains}{List of chains}
}
\value{
Deduplicated chains
}
\description{
Deduplicate chains (v3)
}
\keyword{internal}
