% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_detection.R
\name{detect_all_chains_v3}
\alias{detect_all_chains_v3}
\title{Detect all possible chains in the graph (v3)}
\usage{
detect_all_chains_v3(
  tlx_graph,
  inter_chr_SVs,
  min_chromosomes = 3,
  min_translocations = 3,
  allow_cycles = TRUE,
  max_path_search = 1000,
  max_search_depth = 10,
  max_adjacency_streak = 3
)
}
\arguments{
\item{tlx_graph}{Complete translocation graph}

\item{inter_chr_SVs}{Inter-chromosomal SVs}

\item{min_chromosomes}{Minimum chromosomes}

\item{min_translocations}{Minimum translocations}

\item{allow_cycles}{Allow cycles}

\item{max_path_search}{Maximum paths to search (default: 1000).
Increased to handle complex chromoplexy patterns with many translocations.}

\item{max_search_depth}{Maximum chain length to explore (default: 10).}
}
\value{
List of detected chains
}
\description{
Adapted to work with the new graph structure that includes
genomic adjacency edges.
}
\keyword{internal}
