% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoanasynthesis_detection.R
\name{detect_cn_gradients}
\alias{detect_cn_gradients}
\title{Detect copy number gradients in chromosomes}
\usage{
detect_cn_gradients(
  CNV.sample,
  min_segments = 5,
  gradient_threshold = 0.4,
  max_region_size = 1e+07,
  cn_segment_window_extension = 20,
  cn_window_step_ratio = 4
)
}
\description{
Identifies regions with gradual CN increases or high instability characteristic of chromoanasynthesis.
}
\keyword{internal}
