% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoanasynthesis_detection.R
\name{detect_linked_sv_chains}
\alias{detect_linked_sv_chains}
\title{Detect chains of linked SVs (template switching events)}
\usage{
detect_linked_sv_chains(sv_df, max_gap = 2000)
}
\description{
Detect chains of linked SVs (template switching events)
}
\keyword{internal}
