% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_classifier.R
\name{detect_mechanism_overlaps}
\alias{detect_mechanism_overlaps}
\title{Detect overlapping mechanisms using GenomicRanges}
\usage{
detect_mechanism_overlaps(mechanism_locations, overlap_threshold)
}
\arguments{
\item{mechanism_locations}{Data frame from extract_mechanism_locations}

\item{overlap_threshold}{Minimum overlap in bp}
}
\value{
List of overlapping mechanism pairs
}
\description{
optimized version using interval trees
}
\keyword{internal}
