% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_diagnostics.R
\name{diagnose_single_chain}
\alias{diagnose_single_chain}
\title{Diagnose a single chain}
\usage{
diagnose_single_chain(chain_row, result)
}
\description{
Diagnose a single chain
}
\keyword{internal}
