% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_readers.R
\name{.add_chr_prefix_cnvs}
\alias{.add_chr_prefix_cnvs}
\title{Internal: Add "chr" prefix to chromosome names for CNVsegs object}
\usage{
.add_chr_prefix_cnvs(cnvs)
}
\description{
Internal: Add "chr" prefix to chromosome names for CNVsegs object
}
\keyword{internal}
