% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_readers.R
\name{.add_chr_prefix_svs}
\alias{.add_chr_prefix_svs}
\title{Internal: Add "chr" prefix to chromosome names for SVs object}
\usage{
.add_chr_prefix_svs(svs)
}
\description{
Internal: Add "chr" prefix to chromosome names for SVs object
}
\keyword{internal}
