% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_chainfinder.R
\name{.cf_build_graph}
\alias{.cf_build_graph}
\title{Build graph with statistical adjacency model}
\usage{
.cf_build_graph(
  inter_chr_svs,
  deletion_bridges,
  mu_estimates,
  pvalue_threshold,
  max_distance
)
}
\description{
Build graph with statistical adjacency model
}
\keyword{internal}
