% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_chainfinder.R
\name{.cf_calculate_mu}
\alias{.cf_calculate_mu}
\title{Calculate local rearrangement rate (mu)}
\usage{
.cf_calculate_mu(sv_df, chr_sizes, use_chromosome_specific = TRUE)
}
\description{
Calculate local rearrangement rate (mu)
}
\keyword{internal}
