% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_chainfinder.R
\name{.cf_evaluate_chain}
\alias{.cf_evaluate_chain}
\title{Evaluate a chain}
\usage{
.cf_evaluate_chain(chain, inter_chr_svs, cnv_df, graph, mu_estimates)
}
\description{
Evaluate a chain
}
\keyword{internal}
