% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_chainfinder.R
\name{.cf_find_chains}
\alias{.cf_find_chains}
\title{Find chains using DFS}
\usage{
.cf_find_chains(
  graph,
  min_chromosomes,
  min_translocations,
  max_path_search,
  max_search_depth
)
}
\description{
Find chains using DFS
}
\keyword{internal}
