% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_chainfinder.R
\name{.cf_find_deletion_bridges}
\alias{.cf_find_deletion_bridges}
\title{Find deletion bridges}
\usage{
.cf_find_deletion_bridges(inter_chr_svs, cnv_df, max_distance)
}
\description{
Find deletion bridges
}
\keyword{internal}
