% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_chainfinder.R
\name{.cf_get_chromosome_sizes}
\alias{.cf_get_chromosome_sizes}
\title{Get chromosome sizes}
\usage{
.cf_get_chromosome_sizes(genome = "hg19")
}
\description{
Get chromosome sizes
}
\keyword{internal}
