% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parsers.R
\name{.detect_cnv_caller}
\alias{.detect_cnv_caller}
\title{Detect CNV caller from VCF (VariantAnnotation)}
\usage{
.detect_cnv_caller(vcf)
}
\description{
Detect CNV caller from VCF (VariantAnnotation)
}
\keyword{internal}
