% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parsers.R
\name{.detect_cnv_caller_vcfr}
\alias{.detect_cnv_caller_vcfr}
\title{Detect CNV caller from VCF (vcfR)}
\usage{
.detect_cnv_caller_vcfr(vcf)
}
\description{
Detect CNV caller from VCF (vcfR)
}
\keyword{internal}
