% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parsers.R
\name{.detect_sv_caller}
\alias{.detect_sv_caller}
\title{VCF Parser Functions for Different SV Callers
Internal functions for parsing VCF files from various SV and CNV callers}
\usage{
.detect_sv_caller(vcf)
}
\description{
VCF Parser Functions for Different SV Callers
Internal functions for parsing VCF files from various SV and CNV callers
}
\keyword{(VariantAnnotation)}
\keyword{Detect}
\keyword{SV}
\keyword{VCF}
\keyword{caller}
\keyword{from}
\keyword{internal}
