% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parsers.R
\name{.detect_sv_caller_vcfr}
\alias{.detect_sv_caller_vcfr}
\title{Detect SV caller from VCF (vcfR)}
\usage{
.detect_sv_caller_vcfr(vcf)
}
\description{
Detect SV caller from VCF (vcfR)
}
\keyword{internal}
