% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_variant_annotation_support.R
\name{.infer_sv_type_from_strands}
\alias{.infer_sv_type_from_strands}
\title{Infer SV type from strand orientations}
\usage{
.infer_sv_type_from_strands(chrom1, chrom2, strand1, strand2)
}
\arguments{
\item{chrom1}{First chromosome}

\item{chrom2}{Second chromosome}

\item{strand1}{First strand}

\item{strand2}{Second strand}
}
\value{
Inferred SV type
}
\description{
Infer SV type from strand orientations
}
\keyword{internal}
