% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_criteria.R
\name{.ks_test_exp_mc}
\alias{.ks_test_exp_mc}
\title{Internal Monte Carlo KS test for Exponential Distribution
Corrects for parameter estimation bias (Lilliefors-style)}
\usage{
.ks_test_exp_mc(x, n_sim = 1000)
}
\description{
Internal Monte Carlo KS test for Exponential Distribution
Corrects for parameter estimation bias (Lilliefors-style)
}
\keyword{internal}
