% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_readers.R
\name{.merge_adjacent_segments}
\alias{.merge_adjacent_segments}
\title{Internal: Merge adjacent CNV segments with same CN}
\usage{
.merge_adjacent_segments(df)
}
\description{
Internal: Merge adjacent CNV segments with same CN
}
\keyword{internal}
