% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parsers.R
\name{.parse_delly_sv}
\alias{.parse_delly_sv}
\title{Parse Delly SV VCF (VariantAnnotation)}
\usage{
.parse_delly_sv(vcf, info, ranges)
}
\description{
Parse Delly SV VCF (VariantAnnotation)
}
\keyword{internal}
