% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parsers.R
\name{.parse_generic_sv}
\alias{.parse_generic_sv}
\title{Parse generic SV VCF (VariantAnnotation)}
\usage{
.parse_generic_sv(vcf, info, ranges)
}
\description{
Parse generic SV VCF (VariantAnnotation)
}
\keyword{internal}
