% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parsers.R
\name{.parse_gridss_sv}
\alias{.parse_gridss_sv}
\title{Parse GRIDSS SV VCF (VariantAnnotation)}
\usage{
.parse_gridss_sv(vcf, info, ranges)
}
\description{
Parse GRIDSS SV VCF (VariantAnnotation)
}
\keyword{internal}
