% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parsers.R
\name{.parse_lumpy_sv}
\alias{.parse_lumpy_sv}
\title{Parse LUMPY SV VCF (VariantAnnotation)}
\usage{
.parse_lumpy_sv(vcf, info, ranges)
}
\description{
Parse LUMPY SV VCF (VariantAnnotation)
}
\keyword{internal}
