% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parsers.R
\name{.parse_manta_sv}
\alias{.parse_manta_sv}
\title{Parse Manta SV VCF (VariantAnnotation)}
\usage{
.parse_manta_sv(vcf, info, ranges)
}
\description{
Parse Manta SV VCF (VariantAnnotation)
}
\keyword{internal}
