% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parsers.R
\name{.parse_sniffles_sv}
\alias{.parse_sniffles_sv}
\title{Parse Sniffles SV VCF (VariantAnnotation)}
\usage{
.parse_sniffles_sv(vcf, info, ranges)
}
\description{
Parse Sniffles SV VCF (VariantAnnotation)
}
\keyword{internal}
