% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_readers.R
\name{.read_cnv_vcf_vcfr}
\alias{.read_cnv_vcf_vcfr}
\title{Internal function: Read CNV VCF using vcfR}
\usage{
.read_cnv_vcf_vcfr(vcf_file, caller, cn_field, merge_adjacent, sample_name)
}
\description{
Internal function: Read CNV VCF using vcfR
}
\keyword{internal}
