% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_readers.R
\name{.read_sv_vcf_vcfr}
\alias{.read_sv_vcf_vcfr}
\title{Internal function: Read SV VCF using vcfR}
\usage{
.read_sv_vcf_vcfr(vcf_file, caller, min_sv_size, include_tra, sample_name)
}
\description{
Internal function: Read SV VCF using vcfR
}
\keyword{internal}
