% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_readers.R
\name{.read_unpaired_bnd_from_vcf}
\alias{.read_unpaired_bnd_from_vcf}
\title{Internal function: parse single-record BND variants from VCF ALT notation}
\usage{
.read_unpaired_bnd_from_vcf(vcf_file)
}
\description{
Internal function: parse single-record BND variants from VCF ALT notation
}
\keyword{internal}
