% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_statistics.R
\name{estimate_local_rearrangement_rate}
\alias{estimate_local_rearrangement_rate}
\title{Estimate local rearrangement rate}
\usage{
estimate_local_rearrangement_rate(chromosomes, total_SVs, genome_size)
}
\arguments{
\item{chromosomes}{Chromosomes in chain}

\item{total_SVs}{Total number of SVs in sample}

\item{genome_size}{Genome size}
}
\value{
Local rearrangement rate (mu)
}
\description{
Calculates the expected rate of rearrangements per megabase
for the chromosomes involved in the chain.
}
\keyword{internal}
