% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoanasynthesis_detection.R
\name{evaluate_chromoanasynthesis_region}
\alias{evaluate_chromoanasynthesis_region}
\title{Evaluate a chromoanasynthesis region}
\usage{
evaluate_chromoanasynthesis_region(
  region,
  SV.sample,
  CNV.sample,
  topology,
  min_tandem_dups = 3
)
}
\description{
Evaluate a chromoanasynthesis region
}
\keyword{internal}
