% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_statistics.R
\name{evaluate_deletion_bridge_enrichment_v2}
\alias{evaluate_deletion_bridge_enrichment_v2}
\title{Evaluate deletion bridge enrichment in chain}
\usage{
evaluate_deletion_bridge_enrichment_v2(chain, deletion_bridges)
}
\arguments{
\item{chain}{Chain object}

\item{deletion_bridges}{All deletion bridges}
}
\value{
Enrichment metrics
}
\description{
Evaluate deletion bridge enrichment in chain
}
\keyword{internal}
