% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_detection.R
\name{evaluate_edge_composition_v3}
\alias{evaluate_edge_composition_v3}
\title{Evaluate edge type composition (v3)}
\usage{
evaluate_edge_composition_v3(chain)
}
\description{
Evaluate edge type composition (v3)
}
\keyword{internal}
