% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_statistics.R
\name{evaluate_local_cn_changes_v2}
\alias{evaluate_local_cn_changes_v2}
\title{Evaluate local CN changes near breakpoints}
\usage{
evaluate_local_cn_changes_v2(breakpoints, CNV.sample, window_size = 5e+06)
}
\arguments{
\item{breakpoints}{List of breakpoints}

\item{CNV.sample}{CNV data}

\item{window_size}{Window size for local analysis}
}
\value{
Local CN change score (0-1, higher = more changes)
}
\description{
Evaluate local CN changes near breakpoints
}
\keyword{internal}
