% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint_sequence_analysis.R
\name{extract_breakpoint_sequences}
\alias{extract_breakpoint_sequences}
\title{Extract sequences around breakpoints}
\usage{
extract_breakpoint_sequences(sv_df, genome, flank_size)
}
\arguments{
\item{sv_df}{Data frame with SV information}

\item{genome}{Reference genome}

\item{flank_size}{Flanking sequence size}
}
\value{
Data frame with extracted sequences
}
\description{
Extract sequences around breakpoints
}
\keyword{internal}
