% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_statistics.R
\name{extract_breakpoints_v2}
\alias{extract_breakpoints_v2}
\title{Extract breakpoints from SVs}
\usage{
extract_breakpoints_v2(chain_SVs)
}
\arguments{
\item{chain_SVs}{SV data frame}
}
\value{
List of breakpoints
}
\description{
Extract breakpoints from SVs
}
\keyword{internal}
