% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos_visualization.R
\name{extract_chromoanagenesis_regions}
\alias{extract_chromoanagenesis_regions}
\title{Extract chromoanagenesis regions from result}
\usage{
extract_chromoanagenesis_regions(chromoanagenesis_result, mechanisms)
}
\description{
Extract chromoanagenesis regions from result
}
\keyword{internal}
