% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos_visualization.R
\name{extract_chromoplexy_regions}
\alias{extract_chromoplexy_regions}
\title{Extract chromoplexy regions from chain details}
\usage{
extract_chromoplexy_regions(chain_details)
}
\description{
Extract chromoplexy regions from chain details
}
\keyword{internal}
