% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_classifier.R
\name{extract_mechanism_locations}
\alias{extract_mechanism_locations}
\title{Extract genomic locations of each mechanism}
\usage{
extract_mechanism_locations(chromoanagenesis_result, min_confidence)
}
\arguments{
\item{chromoanagenesis_result}{Chromoanagenesis result object}

\item{min_confidence}{Minimum confidence threshold}
}
\value{
Data frame with mechanism locations
}
\description{
Extract genomic locations of each mechanism
}
\keyword{internal}
