% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint_sequence_analysis.R
\name{extract_sequence_region}
\alias{extract_sequence_region}
\title{Extract sequence from genome region}
\usage{
extract_sequence_region(genome, chrom, start, end, is_bsgenome)
}
\arguments{
\item{genome}{Genome object or file path}

\item{chrom}{Chromosome name}

\item{start}{Start position}

\item{end}{End position}

\item{is_bsgenome}{Whether genome is BSgenome object}
}
\value{
Sequence string
}
\description{
Extract sequence from genome region
}
\keyword{internal}
