% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnome_style_visualization.R
\name{filter_sv_to_region}
\alias{filter_sv_to_region}
\title{Filter SVs to a specific region}
\usage{
filter_sv_to_region(sv_data, chrom, start_pos, end_pos)
}
\description{
Filter SVs to a specific region
}
\keyword{internal}
