% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoanasynthesis_detection.R
\name{find_gradient_windows}
\alias{find_gradient_windows}
\title{Find windows with CN gradients using sliding window approach}
\usage{
find_gradient_windows(
  chr_cnv,
  min_segments = 5,
  gradient_threshold = 0.4,
  max_size = 1e+07,
  cn_segment_window_extension = 20,
  cn_window_step_ratio = 4
)
}
\description{
Find windows with CN gradients using sliding window approach
}
\keyword{internal}
