% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint_sequence_analysis.R
\name{find_microhomology}
\alias{find_microhomology}
\title{Internal: Correctly compare microhomology}
\usage{
find_microhomology(
  s1_kept,
  s2_kept,
  s1_disc,
  s2_disc,
  strand1,
  strand2,
  min_length,
  max_length
)
}
\description{
Internal: Correctly compare microhomology
}
\keyword{internal}
