% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_detection.R
\name{find_nearby_deletions_v3}
\alias{find_nearby_deletions_v3}
\title{Find deletions near a breakpoint (v3)}
\usage{
find_nearby_deletions_v3(chrom, pos, CNV.sample, max_distance)
}
\arguments{
\item{chrom}{Chromosome}

\item{pos}{Position}

\item{CNV.sample}{CNV data}

\item{max_distance}{Maximum search distance}
}
\value{
List of nearby deletions
}
\description{
Find deletions near a breakpoint (v3)
}
\keyword{internal}
