% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_statistics.R
\name{get_genome_size}
\alias{get_genome_size}
\title{Get genome size for reference genome}
\usage{
get_genome_size(genome = "hg19")
}
\arguments{
\item{genome}{Genome build ("hg19" or "hg38")}
}
\value{
Total genome size in bp
}
\description{
Get genome size for reference genome
}
\keyword{internal}
