% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos_visualization.R
\name{merge_sv_with_regions}
\alias{merge_sv_with_regions}
\title{Merge SV data with chromoanagenesis regions}
\usage{
merge_sv_with_regions(sv_data, ca_regions)
}
\description{
Merge SV data with chromoanagenesis regions
}
\keyword{internal}
