% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_analysis.R
\docType{class}
\name{OncoImplexusCohort-class}
\alias{OncoImplexusCohort-class}
\title{Class OncoImplexusCohort}
\description{
A class to represent a cohort of samples with chromoanagenesis analysis results
and associated clinical data.
}
\section{Slots}{

\describe{
\item{\code{results}}{A list of chromoanagenesis result objects (one per sample)}

\item{\code{clinical}}{A data frame containing clinical metadata (rownames must match sample IDs)}

\item{\code{sample_ids}}{A character vector of sample IDs}
}}

