% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint_sequence_visualization.R
\name{plot_breakpoint_report}
\alias{plot_breakpoint_report}
\title{Generate a combined breakpoint sequence analysis report plot}
\usage{
plot_breakpoint_report(repair_data, SV.sample)
}
\arguments{
\item{repair_data}{Result object from analyze_breakpoint_sequences()}

\item{SV.sample}{Original SV data}
}
\value{
A grid arrangement of plots
}
\description{
Generate a combined breakpoint sequence analysis report plot
}
