% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoanasynthesis_visualization.R
\name{plot_chromoanasynthesis}
\alias{plot_chromoanasynthesis}
\title{Plot chromoanasynthesis regions}
\usage{
plot_chromoanasynthesis(
  chromoanasynthesis_result,
  region_id = NULL,
  sample_name = "",
  genome = "hg19"
)
}
\arguments{
\item{chromoanasynthesis_result}{Result from detect_chromoanasynthesis()}

\item{region_id}{Optional: specific region ID to plot (default: plot all)}

\item{sample_name}{Sample name for plot title}

\item{genome}{Reference genome ("hg19" or "hg38")}
}
\value{
A ggplot object or list of ggplot objects
}
\description{
Visualizes chromoanasynthesis regions with CN gradients and tandem duplications.
}
