% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoanasynthesis_visualization.R
\name{plot_chromoanasynthesis_heatmap}
\alias{plot_chromoanasynthesis_heatmap}
\title{Create gradient heatmap for chromoanasynthesis}
\usage{
plot_chromoanasynthesis_heatmap(
  chromoanasynthesis_result,
  region_id,
  sample_name = ""
)
}
\arguments{
\item{chromoanasynthesis_result}{Result from detect_chromoanasynthesis()}

\item{region_id}{Region ID to plot}

\item{sample_name}{Sample name}
}
\value{
ggplot object
}
\description{
Shows CN gradient as a heatmap along chromosome.
}
