% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_visualization.R
\name{plot_chromoplexy_circular}
\alias{plot_chromoplexy_circular}
\title{Create circular plot of chromoplexy chain}
\usage{
plot_chromoplexy_circular(chromoplexy_result, chain_id, sample_name = "")
}
\arguments{
\item{chromoplexy_result}{Result from detect_chromoplexy()}

\item{chain_id}{Chain ID to plot}

\item{sample_name}{Sample name}
}
\value{
A ggplot object
}
\description{
For chains that form cycles, create a circular visualization.
}
