% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_comparison.R
\name{plot_chromoplexy_method_comparison}
\alias{plot_chromoplexy_method_comparison}
\title{Plot comparison of chromoplexy methods}
\usage{
plot_chromoplexy_method_comparison(comparison_df)
}
\arguments{
\item{comparison_df}{Summary data frame from create_comparison_summary_df}
}
\value{
ggplot object
}
\description{
Plot comparison of chromoplexy methods
}
