% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoanasynthesis_visualization.R
\name{plot_cn_gradient_scatter}
\alias{plot_cn_gradient_scatter}
\title{Plot CN gradient scatter plot}
\usage{
plot_cn_gradient_scatter(
  chromoanasynthesis_result,
  region_id,
  sample_name = ""
)
}
\arguments{
\item{chromoanasynthesis_result}{Result from detect_chromoanasynthesis()}

\item{region_id}{Region ID to plot}

\item{sample_name}{Sample name}
}
\value{
ggplot object
}
\description{
Shows correlation between segment order and copy number.
}
