% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_classifier_visualization.R
\name{plot_complexity_breakdown}
\alias{plot_complexity_breakdown}
\title{Plot complexity components breakdown}
\usage{
plot_complexity_breakdown(mixed_mechanisms_result, sample_name = "")
}
\arguments{
\item{mixed_mechanisms_result}{Result from classify_mixed_mechanisms()}

\item{sample_name}{Sample name for plot title}
}
\value{
A ggplot object
}
\description{
Creates a bar plot showing the contribution of different components
to the overall complexity score.
}
