% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_fusion.R
\name{plot_fusion_repair_context}
\alias{plot_fusion_repair_context}
\title{Plot Fusion Repair Context (ggplot2 version)}
\usage{
plot_fusion_repair_context(results, fusion_name, gene_granges)
}
\arguments{
\item{results}{Analysis results.}

\item{fusion_name}{Fusion name.}

\item{gene_granges}{GRanges object with symbols.}
}
\description{
Plot Fusion Repair Context (ggplot2 version)
}
