% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_dashboard_visualization.R
\name{plot_genome_ideogram}
\alias{plot_genome_ideogram}
\title{Plot genome ideogram with chromoanagenesis events}
\usage{
plot_genome_ideogram(
  chromoanagenesis_result,
  sample_name = "",
  genome = "hg19"
)
}
\arguments{
\item{chromoanagenesis_result}{Chromoanagenesis results}

\item{sample_name}{Sample name for subtitle}

\item{genome}{Reference genome ("hg19" or "hg38")}
}
\value{
ggplot2 object
}
\description{
Creates a karyogram-style visualization showing all chromosomes
with chromoanagenesis events highlighted.
}
