% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_dashboard_visualization.R
\name{plot_mechanism_distribution}
\alias{plot_mechanism_distribution}
\title{Plot mechanism distribution summary}
\usage{
plot_mechanism_distribution(chromoanagenesis_result)
}
\arguments{
\item{chromoanagenesis_result}{Chromoanagenesis results}
}
\value{
ggplot2 object
}
\description{
Creates a summary visualization of mechanism proportions.
}
