% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_classifier_visualization.R
\name{plot_mechanism_dominance}
\alias{plot_mechanism_dominance}
\title{Plot mechanism dominance pie chart}
\usage{
plot_mechanism_dominance(mixed_mechanisms_result, sample_name = "")
}
\arguments{
\item{mixed_mechanisms_result}{Result from classify_mixed_mechanisms()}

\item{sample_name}{Sample name for plot title}
}
\value{
A ggplot object
}
\description{
Creates a pie chart showing the relative proportions of different
chromoanagenesis mechanisms.
}
