% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_cohort.R
\name{plot_mechanism_prevalence}
\alias{plot_mechanism_prevalence}
\title{Plot Mechanism Prevalence in a Cohort}
\usage{
plot_mechanism_prevalence(
  summary_df,
  title = "Prevalence of Chromoanagenesis Mechanisms"
)
}
\arguments{
\item{summary_df}{Data frame from summarize_cohort_results().}

\item{title}{Plot title.}
}
\value{
A ggplot object.
}
\description{
Creates a publication-quality bar plot showing the percentage of samples
in the cohort affected by each chromoanagenesis mechanism.
}
