% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint_sequence_visualization.R
\name{plot_microhomology_distribution}
\alias{plot_microhomology_distribution}
\title{Plot microhomology length distribution}
\usage{
plot_microhomology_distribution(repair_data)
}
\arguments{
\item{repair_data}{Result object from analyze_breakpoint_sequences()}
}
\value{
A ggplot2 object
}
\description{
Plot microhomology length distribution
}
