% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnome_style_visualization.R
\name{plot_region_cn}
\alias{plot_region_cn}
\title{Plot region copy number profile}
\usage{
plot_region_cn(cnv_data, start_pos, end_pos)
}
\description{
Plot region copy number profile
}
\keyword{internal}
