% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnome_style_visualization.R
\name{plot_region_ideogram}
\alias{plot_region_ideogram}
\title{Plot region ideogram}
\usage{
plot_region_ideogram(chrom, start_pos, end_pos)
}
\description{
Plot region ideogram
}
\keyword{internal}
