% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint_sequence_visualization.R
\name{plot_repair_by_svtype}
\alias{plot_repair_by_svtype}
\title{Plot repair mechanisms by SV type}
\usage{
plot_repair_by_svtype(repair_data, SV.sample)
}
\arguments{
\item{repair_data}{Result object from analyze_breakpoint_sequences()}

\item{SV.sample}{Original SV data used for analysis}
}
\value{
A ggplot2 object
}
\description{
Creates a stacked bar plot showing the distribution of DNA repair mechanisms
for each structural variant type.
}
