% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint_sequence_visualization.R
\name{plot_repair_mechanisms}
\alias{plot_repair_mechanisms}
\title{Plot inferred repair mechanisms distribution}
\usage{
plot_repair_mechanisms(repair_data)
}
\arguments{
\item{repair_data}{Result object from analyze_breakpoint_sequences()}
}
\value{
A ggplot2 object
}
\description{
Plot inferred repair mechanisms distribution
}
