% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_analysis.R
\name{plot_survival_forest}
\alias{plot_survival_forest}
\title{Plot a Cox regression forest plot}
\usage{
plot_survival_forest(
  survival_result,
  model = c("multivariable", "univariable"),
  top_n = 20
)
}
\arguments{
\item{survival_result}{Result from \code{run_chromoanagenesis_survival()}.}

\item{model}{"multivariable" or "univariable".}

\item{top_n}{Number of features to show.}
}
\value{
A ggplot object.
}
\description{
Plot a Cox regression forest plot
}
