% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_analysis.R
\name{plot_survival_km}
\alias{plot_survival_km}
\title{Plot a Kaplan-Meier curve for one feature}
\usage{
plot_survival_km(survival_result, feature)
}
\arguments{
\item{survival_result}{Result from \code{run_chromoanagenesis_survival()}.}

\item{feature}{Feature name present in \code{survival_result$km_group_data}.}
}
\value{
A ggplot object.
}
\description{
Plot a Kaplan-Meier curve for one feature
}
