% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_dashboard_visualization.R
\name{plot_sv_density_heatmap}
\alias{plot_sv_density_heatmap}
\title{Plot SV density heatmap by chromosome and type}
\usage{
plot_sv_density_heatmap(SV.sample)
}
\arguments{
\item{SV.sample}{SV data (SVs object or data frame)}
}
\value{
ggplot2 object
}
\description{
Creates a heatmap showing the density of different SV types
across chromosomes.
}
