% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_visualization.R
\name{prepare_circular_arcs}
\alias{prepare_circular_arcs}
\title{Prepare arc data for circular plot}
\usage{
prepare_circular_arcs(chain_SVs, chr_positions)
}
\arguments{
\item{chain_SVs}{SVs in chain}

\item{chr_positions}{Chromosome positions in circular layout}
}
\value{
Data frame with arc coordinates
}
\description{
Prepare arc data for circular plot
}
\keyword{internal}
