% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_visualization.R
\name{prepare_translocation_arcs}
\alias{prepare_translocation_arcs}
\title{Prepare translocation arc data for plotting}
\usage{
prepare_translocation_arcs(chain_SVs, chr_info)
}
\arguments{
\item{chain_SVs}{SVs in chain}

\item{chr_info}{Chromosome position info}
}
\value{
Data frame with arc coordinates
}
\description{
Prepare translocation arc data for plotting
}
\keyword{internal}
