% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoanagenesis_integrated.R
\name{print.chromoanagenesis}
\alias{print.chromoanagenesis}
\title{Print method for chromoanagenesis results}
\usage{
\method{print}{chromoanagenesis}(x, ...)
}
\arguments{
\item{x}{Chromoanagenesis result object}

\item{...}{Additional arguments}
}
\description{
Print method for chromoanagenesis results
}
