% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_chainfinder.R
\name{print.chromoplexy_chainfinder}
\alias{print.chromoplexy_chainfinder}
\title{Print method}
\usage{
\method{print}{chromoplexy_chainfinder}(x, ...)
}
\arguments{
\item{x}{A chromoplexy_chainfinder result object}

\item{...}{Additional arguments, currently unused}
}
\value{
Invisibly returns \code{x}
}
\description{
Print method
}
