% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_classifier.R
\name{print.mixed_mechanisms}
\alias{print.mixed_mechanisms}
\title{Print method for mixed mechanisms classification}
\usage{
\method{print}{mixed_mechanisms}(x, ...)
}
\arguments{
\item{x}{Mixed mechanisms result object}

\item{...}{Additional arguments}
}
\description{
Print method for mixed mechanisms classification
}
