% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoplexy_diagnostics.R
\name{print_chromoplexy_thresholds}
\alias{print_chromoplexy_thresholds}
\title{Print classification thresholds}
\usage{
print_chromoplexy_thresholds(
  use_statistical_testing = TRUE,
  fdr_threshold = 0.01
)
}
\arguments{
\item{use_statistical_testing}{Whether statistical testing is enabled}

\item{fdr_threshold}{FDR threshold if using statistical testing}
}
\description{
Shows the thresholds used for chromoplexy classification.
}
