% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoanasynthesis_detection.R
\name{remove_overlapping_windows}
\alias{remove_overlapping_windows}
\title{Remove redundant overlapping windows}
\usage{
remove_overlapping_windows(windows, overlap_cutoff = 0.8)
}
\description{
Remove redundant overlapping windows
}
\keyword{internal}
