% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint_sequence_analysis.R
\name{reverse_complement}
\alias{reverse_complement}
\title{Reverse complement of DNA sequence}
\usage{
reverse_complement(seq)
}
\arguments{
\item{seq}{DNA sequence string}
}
\value{
Reverse complement
}
\description{
Reverse complement of DNA sequence
}
\keyword{internal}
