% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoanagenesis_integrated.R
\name{summary.chromoanagenesis}
\alias{summary.chromoanagenesis}
\title{Summary method for chromoanagenesis results}
\usage{
\method{summary}{chromoanagenesis}(object, ...)
}
\arguments{
\item{object}{Chromoanagenesis result object}

\item{...}{Additional arguments}
}
\description{
Summary method for chromoanagenesis results
}
