% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoanasynthesis_detection.R
\name{summary.chromoanasynthesis}
\alias{summary.chromoanasynthesis}
\title{Summary method for chromoanasynthesis results}
\usage{
\method{summary}{chromoanasynthesis}(object, ...)
}
\arguments{
\item{object}{Chromoanasynthesis result object}

\item{...}{Additional arguments}
}
\description{
Summary method for chromoanasynthesis results
}
