% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromoanasynthesis_detection.R,
%   R/chromoanasynthesis_topology.R
\name{detect_linked_sv_chains}
\alias{detect_linked_sv_chains}
\title{Detect chains of linked SVs (template switching events)}
\usage{
detect_linked_sv_chains(sv_df, max_gap = 2000)

detect_linked_sv_chains(sv_df, max_gap = 2000)
}
\arguments{
\item{sv_df}{Data frame of SVs in the region}

\item{max_gap}{Maximum distance between linked breakpoints (default: 2000bp)}
}
\value{
List of chains (vectors of SV indices)
}
\description{
Identifies chains of structural variants where the end of one SV
is close to the start of another, suggesting serial template switching.
}
\keyword{internal}
